python/cli-payments/requirements.txt
axonfi>=0.1.0
python-dotenv>=1.0.0
